refactor: поменял папку для GameView
