+++ src/views/GameView/GameView.module.css
@@ -1,0 +1,69 @@
+.container {
+  position: relative;
+  width: 100vw;
+  height: 100vh;
+  overflow: hidden;
+}
+
+.saveBtn {
+  position: fixed;
+  bottom: 16px;
+  right: 16px;
+  z-index: 10000;
+  padding: 8px 20px;
+  font-size: 16px;
+  background: #dcbb9a;
+  color: #2e2c2c;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+}
+
+.loadingOverlay {
+  position: fixed;
+  inset: 0;
+  z-index: 9999;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  background: #1a1a2e;
+  color: #e6f0ff;
+  font-family: "Press Start 2P", system-ui, monospace;
+  gap: 24px;
+}
+
+.loadingTitle {
+  font-size: 28px;
+  color: honeydew;
+}
+
+.loadingStage {
+  font-size: 12px;
+  color: #dcbb9a;
+}
+
+.progress {
+  appearance: none;
+  -webkit-appearance: none;
+  width: 300px;
+  height: 20px;
+  border: 2px solid #dcbb9a;
+  border-radius: 4px;
+  background: #2e2c2c;
+}
+
+.progress::-webkit-progress-bar {
+  background: #2e2c2c;
+  border-radius: 4px;
+}
+
+.progress::-webkit-progress-value {
+  background: #dcbb9a;
+  border-radius: 2px;
+}
+
+.progress::-moz-progress-bar {
+  background: #dcbb9a;
+  border-radius: 2px;
+}
